refactor(View): separate into input and output view
